Add a new section 'Guidelines' to CoC. (#23)
After reading https://www.newyorker.com/news/letter-from-silicon-valley/the-lonely-work-of-moderating-hacker-news and finding it quite interesting, I had a look at their guidelines: https://news.ycombinator.com/newsguidelines.html and found the paragraphs in this commit really good. It's kind of an addition to the tl;dr with some easy to understand guidelines on how we'd like people to conduct themselves.