Remove mention of #codeofconduct channel
We should
- remove/archive the #codeofconduct channel.
- remove it from the default channels
- document how to create an 🔒admins- channel

Discussions around the Code of Conduct can be done in 🔒admins- channels.

\#codeofconduct being a default channel has a few downsides. It has been used in bad
faith by people to dogpile other users or threads, derailing any productive
or valuable discussion.

It's also been abused by people trying to broadcast messages to a lot of
people even when asked to no longer comment on incidents or threads,
which then wastes more time of the admins and moderators, when they're
already trying to prioritise the right issues and reduce harm.

It's also been used by mistake by people which then spams 8000
users.